backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
networkx
python-dotenv
websockets
